+++ bin/cli.php
@@ -9,6 +9,7 @@ use Symfony\Component\Console\Application;
 use App\Console\CronNotifications;
 use App\Console\CronQuarantine;
 use App\Console\CronUpdateMapFiles;
+use App\Console\UserAdd;
 
 $application = new Application();
 
@@ -17,5 +18,6 @@ $application = new Application();
 $application->add(new CronNotifications($fileLogger, $syslogLogger));
 $application->add(new CronQuarantine($fileLogger, $syslogLogger));
 $application->add(new CronUpdateMapFiles($fileLogger, $syslogLogger));
+$application->add(new UserAdd($fileLogger, $syslogLogger));
 
 $application->run();